requirements.txt
pdfminer.six==20221105
sentence-transformers==3.2.1
torch==2.1.2